refactor(cli): extract JS-shaped path encoding to pkg/files/encodepath
Move EncodeURL / EncodeURIComponent out of package upload into a small
shared package so download, rm, path, and upload all depend on the same
wire encoder without cmd/pkg layers importing upload only for URLs.

- Add cli/pkg/files/encodepath with tests (former upload/encode*.go).
- Update upload api/uploader, download client, rm, FrontendPath.URLPath.
- Clarify download/list.go is for the download walker, not `files ls`.

Made-with: Cursor
