+++ cli/cmd/ctl/files/ls.go
@@ -125,7 +125,7 @@ func runLs(ctx context.Context, f *cmdutil.Factory, out io.Writer, rawPath strin
 		return err
 	}
 
-	// URLPath uses upload.EncodeURL (same as download/cat/rm/upload) so
+	// URLPath uses encodepath.EncodeURL (same as download/cat/rm/upload) so
 	// filenames with '#', '?', '+', spaces, '!*'()', etc. survive
 	// the trip to the backend. ParseFrontendPath already guarantees that
 	// listing the extend root ("drive/Home" or "drive/Home/") yields a
